// src/app/model/receita-model.ts
export class ReceitaModel {
    codigo: number;
    tipoReceita: string;
    responsavelPagamento: any;
    categoriaReceita: any;
    dataPrevisaoRecebimento: string;
    dataRecebimentoPagamento: string;
    valorPagamento: string;
    quantidadeParcela: number;
    tipoFormaPagamento: any;
    contaBancariaDeposito: any;
    tipoSituacaoPagamento: any;
    tipoPeriodoFinanceiro: any;
}